### 2019-08-02 #1622 @labarba
👋 @ShixiangWang — You provided the archive DOI instead of the repository URL, and our bot cannot build your paper.

Please confirm that this is the correct submission repository:
https://github.com/ropensci/UCSCXenaTools/
