### 2019-08-03 #1622 @ShixiangWang
@labarba Yes, this is the package repository. I followed the instruction from https://github.com/ropensci/software-review/issues/315#issuecomment-515795175 to submit a Zenodo DOI instead of GitHub repository. I am sorry if this operation was wrong.
