### 2019-08-03 #1622 @labarba
@arfon — I'm confused. The author was told to "Submit to JOSS using the Zenodo DOI" after acceptance of his software on rOpenSci. We need the software repo URL, so I think there's some confusion here.
