### 2019-08-05 #1622 @ShixiangWang
@labarba Maybe rOpenSci should update the approval comment template.

Could you help me correct the repository URL?

I still see the Zenodo DOI, but I cannot edit it.

![image](https://user-images.githubusercontent.com/25057508/62433587-94f35400-b767-11e9-9331-29c6f9e7bd5c.png)
