### 2019-08-05 #1622 @arfon
> The author was told to "Submit to JOSS using the Zenodo DOI" after acceptance of his software on rOpenSci. We need the software repo URL, so I think there's some confusion here.

Yes, that's not correct. @ShixiangWang - could you point me to the documentation from rOpenSci that asks for this?
