### 2019-08-05 #1627 @whedon
[REVIEW]: The UCSCXenaTools R package: a toolkit for accessing genomics data from UCSC Xena platform, from cancer multi-omics to single-cell RNA-seq
**Submitting author:** @ShixiangWang (<a href="http://orcid.org/0000-0001-9855-7357">Wang Shixiang</a>)
**Repository:** <a href="https://github.com/ropensci/UCSCXenaTools" target ="_blank">https://github.com/ropensci/UCSCXenaTools</a>
**Version:** v1.2.5
