### 2019-08-05 #1627 @arfon
Accepting this submission as it's already been [accepted by rOpenSci](https://github.com/ropensci/software-review/issues/315)
