### 2019-08-05 #1627 @arfon
@ShixiangWang - your submission is now accepted into JOSS :zap::rocket::boom:
