### 2019-08-05 #1622 @ShixiangWang
@arfon It's from to do list at here https://github.com/ropensci/software-review/issues/315#issuecomment-515795175
