### 2019-08-05 #1622 @noamross
Thanks for pointing this out, @ShixiangWang.  @arfon, we'll fix this in our templates.  Has the submission form changed? I seem to remember that one previously had to put the Zenodo DOI into the form (in addition to the repo URL).
