### 2019-08-05 #1622 @arfon
> Has the submission form changed?

Nope, it's the same as it's always been. 😁
